requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
astroid==3.1.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
isort==5.13.2
magic-filter==1.0.12
mccabe==0.7.0
multidict==6.0.5
numpy==1.26.4
openai==1.17.1
pendulum==3.0.0
pillow==10.2.0
platformdirs==4.2.0
psycopg2==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
pylint==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
requests==2.31.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
time-machine==2.14.1
tomli==2.0.1
tomlkit==0.12.4
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
